requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
python-dotenv==1.0.0
boto3==1.29.3
sentence-transformers==2.2.2
faiss-cpu==1.7.4
numpy==1.24.3
torch==2.1.0
networkx==3.2.1
node2vec==0.4.6
pydantic==1.10.13
langchain==0.0.335
openai==1.3.0
python-multipart==0.0.6
prisma==0.11.0
requests==2.31.0 